[#14] 🩹 fix : profile의 user 엔티티 수정
