+++ src/main/java/com/sparta/eduwithme/domain/profile/ProfileRepository.java
@@ -1,9 +1,7 @@
 package com.sparta.eduwithme.domain.profile;
 
-import com.sparta.eduwithme.domain.profile.entity.User;
+import com.sparta.eduwithme.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
 
 public interface ProfileRepository extends JpaRepository<User, Long> {
 }
