+++ src/main/java/com/sparta/eduwithme/domain/profile/dto/UpdateNicknameRequestDto.java
@@ -7,11 +7,11 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class UpdateNicknameRequestDto {
     private String email;
-    private String nickname; // 사용자로부터 받을 닉네임
+    private String nickName; // 사용자로부터 받을 닉네임
 
     // 인수 생성자
-    public UpdateNicknameRequestDto(String email, String nickname) {
+    public UpdateNicknameRequestDto(String email, String nickName) {
         this.email = email;
-        this.nickname = nickname;
+        this.nickName = nickName;
     }
 }
