+++ src/main/java/com/sparta/eduwithme/domain/profile/dto/UpdateNicknameResponseDto.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class UpdateNicknameResponseDto {
-    private String nickname; // 서버가 반환할 닉네임
+    private String nickName; // 서버가 반환할 닉네임
 
     // 생성자
-    public UpdateNicknameResponseDto(String nickname) {
-        this.nickname = nickname;
+    public UpdateNicknameResponseDto(String nickName) {
+        this.nickName = nickName;
     }
 }
